add "yum_proxy" option to builder
Previously if you wanted to use an HTTP proxy server with a builder, you
had to either specify the proxy in the pkgurl, or use iptables to
intercept HTTP and route it through a transparent proxy.

There is a better way: yum has proxy support built in. When a URL is
specified in the "proxy" option in yum.conf, yum will use this URL as a
proxy server.

Add a "yum_proxy" option in kojid.conf. Kojid will pass it back to mock
during genMockConfig(). Mock will write it into yum.conf in the chroot.