kojivmd: Call out to virt-clone instead of virtinst
Reason is that virtinst is going away as a public library, hopefully
for F17. It wasn't ever really intended to be used publically, and it's
not very good :) Long term it will mostly go away anyways as we move to
libvirt-gobject based libraries.

Unfortunately this change doesn't simplify the koji code any, but it
should be equivalent functionality. I did some local unit testing to
verify it was giving sane results, but I'd appreciate if someone with
an actual koji setup could test.