koji: small patch to fix warnings
After the koji.conf 'pkgurl' attribute was enabled (sigul wants this;
I'll submit a patch for that too, later), one of my scripts broke
because the "Warning: the pkgurl option is obsolete" message is printed
to stdout instead of stderr.  Here's a little patch to print the warning
with the warn() function instead.

I realize some warnings from the cli might be best printed to stdout,
but for ones that might appear when using a computer-readable type of
command like 'list-* --quiet', stderr seems like the better choice.

Thanks-

	John