koji wait-repo patch
Hi,

I've been bitten a few times running "koji wait-repo" that the build I'm
waiting on isn't actually tagged in the right place to get included in the
repo. (Or is blocked, or is typoed etc).

I think koji wait-repo T --build X should actually check that X is the latest
build in T, and warn you if it's not. (Warn not abort in case you're being
clever and running wait-repo at the same time as you build the package)

Here's a patch: