remove the resolve-plugins goal from Maven builds
The resolve-plugins goal is not often used during upstream development, and it
causes unnecessary dependencies to be downloaded and builds to take longer.
The data collected during that step hasn't been used for anything, so it is
safe to remove.  All dependencies downloaded during the build will still be
recorded after the deploy goal is completed.