Allow specifying an 'int' for 'latest'.
This should preserve the pre-existing behavior, where if latest is set
to True, only the latest builds in a tag are returned, and if latest is
set to False, all the builds in a tag are returned.

The new behavior is a kind of middleground, where you can set latest=3,
and only the latest three builds for each package in a tag are returned.

We want this ultimately for mashing rawhide so we can create a downgrade
path for people if new builds break them.
https://bugzilla.redhat.com/show_bug.cgi?id=1082830

Discussed and revised:
https://lists.fedoraproject.org/pipermail/buildsys/2015-February/004541.html