(koji-web) Treat extra-footer as a template.
The code previously would look for a file called extra-footer.html that a koji
admin could override to put some nice extra footer on their koji-web page.
If that file existed, koji-web would read it in verbatim and write it to the
bottom of each served page.

This change enhances that behavior and treats the extra-footer.html file as a
Cheetah template.  This way, a koji admin could specify some conditional logic
in the extra-footer file to display this or that thing depending on the dynamic
contents of the page being served.

Note that this is backwards compatible.  Static html extra-footer files
already out there in the wild will be interpreted as Cheetah templates and
should be included seamlessly.