Don't retry on SSL failures
With the current code, trying to SSL-login with a bad certificate will
just make it look like the client code is hanging.

That's because it tries and tries again, silently, until it reaches it's
maximum retry limit.

But in the case of an SSL error, such as an expired client cert, there's
really no point in retrying.