Add packages to a koji package group as mandatory not default
All the packages in groups that koji uses to build srpms/rpms/live images and other outputs are mandatory but are currently specified as default. Yum historically didn't make a major diffentiation between the two and failed build if there was issues, dnf treats this differently and only classifies mandatory as the only option that would fail the install. The packages in the group should be mandatory because we really do want to ensure they're there. This fixes dnf, but should have no impact on the yum use case.

Signed-off-by: Peter Robinson <pbrobinson@fedoraproject.org>