Allow hiding a user from the frontpage task list.
This adds new query arguments to the taskList hub xmlrpc endpoint, and then
makes use of those arguments in koji-web.  A new optional configuration value
is added for koji-web: `HiddenUser`, which can be used to specify which user
account should be hidden.  This could be useful for deployments that have a
continuous-integration account, the spam from which makes the frontpage
difficult to read.

Unit test cases are also added for some functions of the hub taskList endpoint.

Signed-off-by: Ralph Bean <rbean@redhat.com>