Don't cancel subtask when other fails
In some cases it makes sense to continue with sibling task even if some of
them already failed. E.g. with kernel build it could be of use if submitter
knows for which archs it succeed and for which it fails. Repeated builds
could take a lot of time and resources.

Option is turned on by 'build_arch_can_fail' in kojid.conf and it is turned
on per builder. Note, that it shouldn't be enabled by default as it could
result in unnecessary resource consumption.