pass build information consistently to the postImport callback
Always pass the build infomation as the "build" argument to the
postImport callback. For type='build' and type='image' imports,
this change is additive, and so is unlikely to affect existing
plugins. For type='cg' imports, this change makes the callback
consistent with the other callback types, and the CG import
mechanism is new enough that there is probably no code making
use of it yet.