provide more information to the pre/postRPMSign callback
The existing pre/postRPMSign callback was following the format of
the Task/BuildStateChange callbacks. This isn't necessary for a
one-time operation. This change removes the attribute/old/new
arguments, and replaces them with sigkey, sighash, build, and rpm
arguments. This is more clear, and should make writing handlers
for this callback simpler.