Revert "default krb_rdns to True"
This reverts commit 02128ee5777b560619314cc0220e1ed7ca16b28c.

The second column in cfgmap is an alias and "True" is not a valid alias.
It was mistaken in the reverted commit as a default value, which it is
not. Reverting this commit should fix the following traceback that
koji-gc in Fedora currenty raises:

Traceback (most recent call last):
File "/usr/sbin/koji-gc", line 901, in <module>
options, args = get_options()
File "/usr/sbin/koji-gc", line 159, in get_options
if config.has_option(*alias):
TypeError: has_option() argument after * must be a sequence, not bool