update the 1.11 to 1.12 upgrade schema for BDR
in Fedora we are trying to use BDR to make the koji db more resilient
in order to get the schema to apply I had to make some changes in
order for the schema changes to be applied.
see http://bdr-project.org/docs/stable/ddl-replication-statements.html
for more info there is examples that matched what we are doing at the
bottom of the page  in the section "8.2.3. DDL statements with
restrictions"

Signed-off-by: Dennis Gilmore <dennis@ausil.us>