kojibuilder: Pass mergerepo_c --all for bare mode as well.
The '--koji' mode of mergerepo_c implies --all, and thats what we want
for bare mode as well. The default merge mode is 'repo' which uses the
first found package by name if multiple packages with the same name are
in the multiple repos. This causes chaos with rhel8 repos as GA and
updates content are in the same repo and if it hits an old version of
one package and an update of another that depends on the new version
it blows up.

Signed-off-by: Kevin Fenzi <kevin@scrye.com>