BuildSRPMFromSCMTask: Support auto-selecting a matching specfile name
There are cases where having two files with the .spec suffix in the
same dist-git repository makes sense, such as maintaining a common
dist-git structure for different distributions where the package
name might be different.

This patch would allow the build-system to use the matching spec
file name for the current SCM checkout path if it exists.

Signed-off-by: Stephen Gallagher <sgallagh@redhat.com>