doc: improve hub selinux instructions
Make the Server HowTo SELinux instructions simpler and easier to read.
Describe each step for users to copy-and-paste.

Use semanage and restorecon instead of chcon in order to make the
changes persist across relabling operations.

Only recommend httpd_can_network_connect_db=1 if the user is using a
TCP/IP connection to PostgreSQL. If koji-hub is using a local Unix
socket connection, Apache does not need this setting.