livemedia-creator: pass --nomacboot on non-x86_64
See https://pagure.io/releng/issue/9601 . It seems the default
is 'try to enable Mac booting' and we need to explicitly pass
`--nomacboot` to turn it off. We only support x86_64 Macs for
now, so we don't want to try and make live images bootable on
Mac on any other arch.

Signed-off-by: Adam Williamson <awilliam@redhat.com>