cli: improve --config and --profile help text
Move "--profile" to the top of the argument list so it's easier to see
in --help. Most users should use --profile instead of --config.

Explain what the default --profile value is (it's progname).

Reword the "--config" help text to explain that this overrides a
profile's settings using a separate configuration file.