cli: mention --profile in --config help text
It's easy for new users to confuse the "--config" and "--profile"
settings.

We want to guide users towards using non-default profile names (eg
"stream" or "cbs", etc) where possible. If a user is pointing at a
non-standard configuration file location, they should really use the
--profile option also to keep their configuration sane.

Update the --help text for the --config option to emphasize the
--profile option.