www: fix alignments in channelinfo hosts table
Use the "data-list" classes to style the Hosts table in the channelinfo
page. This makes the header row stand out and easier to read, matching
the rest of the data tables in kojiweb.

This change also fixes a problem in some downstream themes. For example,
Fedora's theme and Brew's theme set "text-align: right" for all tbody.th
elements. These themes do this in order to control the larger parent
"info" table header cells, but it ends up mis-aligning un-classed
sub-tables like this one. With this change, the "list-header" class
left-aligns the header row text, and the "data-list" class left-aligns
the final text.