kojivmd: narrow error handling for missing VMs
lookupByName() could raise libvirt.libvirtError for many different
reasons (libvirt connection problems, etc).

If this kojivmd host does not have this VM available
(VIR_ERR_NO_DOMAIN), and we should log the "VM not available" message
and skip taking the task.

If the error is something else, this is unexpected, and we should raise
it in the logs so the administrator can see it at non-debug log levels.