cli: fix wait-repo message when missing builds
The wait-repo command performs a sanity-check when the user specifies an
NVR with "--build". If there are no builds of the package in the tag at
all, wait-repo will print a warning.

The text of that warning was not accurate. The package itself may be in
the tag (listPackages), but there are no builds of that package
(getLatestBuilds). Correct the warning text to clarify the exact problem
to the user.