test_pages updates
This fixes issues with test_pages that I found while checking
coverage. The environ dict was recycled across each call in the loop,
but those calls can modify it, leaking data from one pass to another.
Also, the mocking for _getServer was not placing the session in the
environ, which was causing the tests to miss some cases.